Disable forceOpacity option for now
It would only cause confusion in its current state because it can
sometimes make elements look bugged even though it's working as
intended. The rest of the time, it would seem like a bug that it isn't
making the underlying element visible since it usually isn't due to
opacity. Replacing this with something that would make it possible to
"tunnel through" to seeing the focused element in ALL cases (not just
opacity) could be a valid option to replace this toggle in the future.